convert crdt tests to java tests
