+++ src/javaharness/java/arcs/crdt/BUILD
@@ -2,7 +2,10 @@ load("@io_bazel_rules_closure//closure:defs.bzl", "closure_js_library")
 
 licenses(["notice"])
 
-package(default_visibility = ["//java/arcs:__subpackages__"])
+package(default_visibility = [
+    "//java/arcs:__subpackages__",
+    "//javatests/arcs:__subpackages__",
+])
 
 load("@com_google_j2cl//build_defs:rules.bzl", "j2cl_library")
 load("@build_bazel_rules_android//android:rules.bzl", "android_library")
