+++ src/javaharness/javatests/arcs/android/impl/CrdtTest.java
@@ -1,43 +1,33 @@
-package arcs.crdt;
+package arcs.android.impl;
 
+import arcs.crdt.CollectionData;
+import arcs.crdt.CollectionOperation;
+import arcs.crdt.CRDTCollection;
+import arcs.crdt.CrdtTestHelper;
+import arcs.crdt.CrdtTestHelper.Data;
+import arcs.crdt.RawCollection;
+import arcs.crdt.VersionMap;
 import java.util.Arrays;
 import java.util.Objects;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.JUnit4;
 
 // The set of tests copied from src/runtime/crdt/tests/crdt-collection-test.ts
 // Please, keep in sync.
 
 @SuppressWarnings("unchecked")
-public class CollectionDataTest {
-  public static class Data implements Referenceable {
-    public final String id;
-
-    public Data(String id) {
-      this.id = id;
-    }
-
-    @Override
-    public String getId() {
-      return id;
-    }
-
-    @Override
-    public boolean equals(Object other) {
-      return this.id.equals(((Data) other).id);
-    }
-
-    @Override
-    public int hashCode() {
-      return this.id.hashCode();
-    }
-  }
-
-  public static void testInitiallyIsEmpty() {
+@RunWith(JUnit4.class)
+public class CrdtTest {
+  @Test
+  public void testInitiallyIsEmpty() {
     // initially is empty
     CRDTCollection<Data> set = new CRDTCollection<>();
     verifySize(set, 0);
   }
 
-  public static void testTwoItemsSameActor() {
+  @Test
+  public void testTwoItemsSameActor() {
     // can add two different items from the same actor
     CRDTCollection<Data> set = new CRDTCollection<>();
     assert set.applyOperation(
@@ -51,7 +41,8 @@ public class CollectionDataTest {
     verifyIds(set, new String[] {"one", "two"});
   }
 
-  public static void testSameValueTwoActors() {
+  @Test
+  public void testSameValueTwoActors() {
     // can add the same value from two actors
     CRDTCollection<Data> set = new CRDTCollection<>();
     assert set.applyOperation(
@@ -64,8 +55,8 @@ public class CollectionDataTest {
     verifyIds(set, new String[] {"one"});
   }
 
-  public static void testRejectAddsNotInSequence() {
-    // rejects add operations not in sequence
+  @Test
+  public void testRejectAddsNotInSequence() {
     CRDTCollection<Data> set = new CRDTCollection<>();
     assert set.applyOperation(
         new CollectionOperation<>(
@@ -81,7 +72,8 @@ public class CollectionDataTest {
             CollectionOperation.Type.ADD, new Data("two"), VersionMap.of("me", 3), "me"));
   }
 
-  public static void testRemoveItem() {
+  @Test
+  public void testRemoveItem() {
     // can remove an item
     CRDTCollection<Data> set = new CRDTCollection<>();
     assert set.applyOperation(
@@ -94,7 +86,8 @@ public class CollectionDataTest {
     verifySize(set, 0);
   }
 
-  public static void testRejectRemoveIfVersionMismatch() {
+  @Test
+  public void testRejectRemoveIfVersionMismatch() {
     // rejects remove operations if version mismatch
     CRDTCollection<Data> set = new CRDTCollection<>();
     assert set.applyOperation(
@@ -108,7 +101,8 @@ public class CollectionDataTest {
             CollectionOperation.Type.REMOVE, new Data("one"), VersionMap.of("me", 0), "me"));
   }
 
-  public static void testRejectRemoveNonexistent() {
+  @Test
+  public void testRejectRemoveNonexistent() {
     // rejects remove value not in collection
     CRDTCollection<Data> set = new CRDTCollection<>();
     assert set.applyOperation(
@@ -119,7 +113,8 @@ public class CollectionDataTest {
             CollectionOperation.Type.REMOVE, new Data("two"), VersionMap.of("me", 1), "me"));
   }
 
-  public static void testRejectRemoveTooOld() {
+  @Test
+  public void testRejectRemoveTooOld() {
     // rejects remove version too old
     CRDTCollection<Data> set = new CRDTCollection<>();
     assert set.applyOperation(
@@ -140,62 +135,11 @@ public class CollectionDataTest {
             CollectionOperation.Type.REMOVE, new Data("two"), VersionMap.of("me", 1), "them"));
   }
 
-  public static void testMergeModels() {
-    // can merge two models
-    CRDTCollection<Data> set1 = new CRDTCollection<>();
-    assert set1.applyOperation(
-        new CollectionOperation<>(
-            CollectionOperation.Type.ADD, new Data("one"), VersionMap.of("me", 1), "me"));
-    assert set1.applyOperation(
-        new CollectionOperation<>(
-            CollectionOperation.Type.ADD, new Data("two"), VersionMap.of("me", 2), "me"));
-    CRDTCollection<Data> set2 = new CRDTCollection<>();
-    assert set2.applyOperation(
-        new CollectionOperation<>(
-            CollectionOperation.Type.ADD, new Data("three"), VersionMap.of("you", 1), "you"));
-    assert set2.applyOperation(
-        new CollectionOperation<>(
-            CollectionOperation.Type.ADD, new Data("one"), VersionMap.of("you", 2), "you"));
-    MergeResult<?> result1 = set1.merge(set2.getData());
-
-    CollectionData<Data> expectedSet1 = new CollectionData<>();
-    expectedSet1.values.put(
-        "one", new VersionedValue<>(new Data("one"), VersionMap.of("me", 1, "you", 2)));
-    expectedSet1.values.put("two", new VersionedValue<>(new Data("two"), VersionMap.of("me", 2)));
-    expectedSet1.values.put(
-        "three", new VersionedValue<>(new Data("three"), VersionMap.of("you", 1)));
-    expectedSet1.version = VersionMap.of("you", 2, "me", 2);
-    CollectionChange<Data> modelChange1 = (CollectionChange<Data>) result1.modelChange;
-    assert modelChange1.changeType == ChangeType.MODEL
-        : "modelChange1.changeType should be ChangeType.MODEL";
-    assert collectionDeepEquals(
-            (CollectionData<Data>) modelChange1.modelPostChange.get(), expectedSet1)
-        : "Unexpected merge model change (1).";
-    assert changeDeepEquals(modelChange1, (CollectionChange<Data>) result1.otherChange)
-        : "modelChange1 must be equal otherChange1";
-
-    // Test removes also work in merge.
-    set1.applyOperation(
-        new CollectionOperation<>(
-            CollectionOperation.Type.REMOVE,
-            new Data("one"),
-            VersionMap.of("me", 2, "you", 2),
-            "me"));
-
-    MergeResult<?> result2 = set1.merge(set2.getData());
-    CollectionData<Data> expectedSet2 = new CollectionData<>();
-    expectedSet2.values.put("two", new VersionedValue<>(new Data("two"), VersionMap.of("me", 2)));
-    expectedSet2.values.put(
-        "three", new VersionedValue<>(new Data("three"), VersionMap.of("you", 1)));
-    expectedSet2.version = VersionMap.of("you", 2, "me", 2);
-    CollectionChange<Data> modelChange2 = (CollectionChange<Data>) result2.modelChange;
-    assert modelChange2.changeType == ChangeType.MODEL
-        : "modelChange2.changeType should be ChangeType.MODEL";
-    assert collectionDeepEquals(
-            (CollectionData<Data>) modelChange2.modelPostChange.get(), expectedSet2)
-        : "Unexpected merge model change (2).";
-    assert changeDeepEquals(modelChange2, (CollectionChange<Data>) result2.otherChange)
-        : "modelChange2 must be equal otherChange2";
+  @Test
+  public void testMergeModels() {
+    // This test cannot be moved, because it heavily uses package private methods and data members.
+    // TODO: refactor?
+    CrdtTestHelper.testMergeModels();
   }
 
   private static void verifySize(CRDTCollection<Data> set, int expectedSize) {
@@ -217,30 +161,6 @@ public class CollectionDataTest {
             + String.join(",", setIds)
             + "]";
   }
-
-  private static boolean collectionDeepEquals(
-      CollectionData<Data> set1, CollectionData<Data> set2) {
-    return Objects.deepEquals(set1.version, set2.version)
-        && Objects.deepEquals(set1.values.keySet(), set2.values.keySet())
-        && set1.values.entrySet().stream()
-            .allMatch(
-                entry -> {
-                  VersionedValue<Data> v2 = set2.values.get(entry.getKey());
-                  return Objects.deepEquals(entry.getValue().value, v2.value)
-                      && Objects.deepEquals(entry.getValue().version, v2.version);
-                });
-  }
-
-  private static boolean changeDeepEquals(
-      CollectionChange<Data> change1, CollectionChange<Data> change2) {
-    assert change1.changeType == ChangeType.MODEL : "unsupported change type comparison";
-    return change1.changeType.equals(change2.changeType)
-        && collectionDeepEquals(
-            (CollectionData<Data>) change1.modelPostChange.get(),
-            (CollectionData<Data>) change2.modelPostChange.get());
-  }
-
-  private CollectionDataTest() {}
 }
 
 // Note: if/when adding more tests to this file, please, also update crdt-collection-test.ts
