+++ src/runtime/crdt/tests/crdt-collection-test.ts
@@ -486,4 +486,4 @@ describe('CRDTCollection', () => {
   });
 });
 
-// Note: if/when adding more tests to this file, please, also update CollectionDataTest.java
+// Note: if/when adding more tests to this file, please, also update CrdtTest.java
